requirements.txt
Flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
pyotp==2.9.0
python-dotenv==1.0.0
email-validator==2.0.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
eventlet==0.36.1
pytest==7.4.3
pytest-flask==1.3.0
